requirements.txt
PyQt6>=6.6.0
pandas>=2.0.0
openpyxl>=3.1.0
gotrue>=2.0.0
python-dotenv>=1.0.0
httpx>=0.26.0
